Added support for verifying the SPI post write
This makes two changes to improve our confidence that we know the exact
state of the SPI post-flash.

Previously when flashing the SPI areas that would not touch any areas
that are not being used. In practice this wasn't a huge issue because
the only parts that weren't touched was the padding around pieces of the
firmware. But if we ever rearange things we might leave stale data lying
around. So I changed the spi write process to write 0xff to previously
unwritten areas. This also has the nice added benefit of turning the
multiple writes into a single one, which doesn't have any practical
benafit, but I like the aesthetics of if :)

I also added a post flash verification step, if there is a flash failure
you likely won't have anything that you can do, but at least there will
be a nice error message with some info that your board might not be
alive after the next reset.