Fixed error on Python3.9 import
As suggested by @frkl Python3.9 seems to have a stdlib regression where
the importlib.resources path function no longer exists. According to the
docs it is eventually deprecated in Python3.11 with the recommended
replacement being as_file(files(...)). These functions don't exist pre
3.9 so I added the importlib-resources package which backports these
changes.

As a sidenote textual (imported by tt-tools-common) requires a much more
recent version of importlib-resources (and metadata) than I need. So I
fixed them to >= a minimum version. This will probably cause problems
when Python inevitably breaks stuff again. So we probably need to test
on a matrix of Python versions (not just oldest and newest) to catch
this in the future.

Signed-off-by: Daniel Rosen <drosen@tenstorrent.com>