Added support for getting the p300 board type
The p300 boards have two firmware packages one for the left chip and
another for the right. Therefore in order to flash the correct fw to the
correct chip we must query the fw for which chip we are looking at. This
is in contrast to the other boards which only require the product ID to
be readback from the PCI config space.

In practice this won't be a hugely problematic change because the BH fw
update already has a FW dependency.

Signed-off-by: Daniel Rosen <3820236+DRvader@users.noreply.github.com>