if FW version is 80.X.Y.Z, consider X the major version
Old FW bundles used to start with 80 and the version format was 80.major.minor.patch.
FW version switched over at major version 18 from 80.18.X.X to 18.X.X.
We need to make an exception when detecting FW versions for the upgrade/downgrade
logic to ignore the 80 and consider the value at index 1 to be the major version.

We can probably revisit and remove this later if we cease to support
FW versions starting with 80.

Signed-off-by: Petra Alexson <palexson@tenstorrent.com>